Moved homeMenuBar component homeMenuBar.js
